+++ apps/builder/src/types/builder/unified.types.ts
@@ -1879,7 +1879,7 @@ export function createDefaultBodyProps(): DivElementProps {
     className: "react-aria-Body",
     style: {
       display: "block",
-      fontFamily: `"Pretendard", "Inter Variable", monospace, system-ui, sans-serif`,
+      fontFamily: `"Pretendard", "Inter Variable", system-ui, sans-serif`,
       overflow: "auto",
     },
   };